lambda_ingest/requirements.txt
boto3
psycopg2-binary
fitz
PyMuPDF
sentence-transformers